Update dotnet project, split up website from converter implementation
