+++ src/dotnet/SudzC.sln
@@ -1,33 +1,61 @@
 ﻿
-Microsoft Visual Studio Solution File, Format Version 11.00
-# Visual Studio 2010
-Project("{E24C65DC-7377-472B-9ABA-BC803B73C61A}") = "Website", "..\Website\", "{838EA5E3-8A94-4D2B-AF1D-121BF7AA2B0E}"
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 2012
+Project("{E24C65DC-7377-472B-9ABA-BC803B73C61A}") = "Website", "http://localhost:57492", "{E7ABC5AC-CB11-48B0-A07E-D36D0F5DB7AE}"
 	ProjectSection(WebsiteProperties) = preProject
-		TargetFrameworkMoniker = ".NETFramework,Version%3Dv4.0"
-		Debug.AspNetCompiler.VirtualPath = "/Website"
-		Debug.AspNetCompiler.PhysicalPath = "..\Website\"
-		Debug.AspNetCompiler.TargetPath = "PrecompiledWeb\Website\"
+		UseIISExpress = "true"
+		TargetFrameworkMoniker = ".NETFramework,Version%3Dv4.5"
+		ProjectReferences = "{d3173d35-982e-4a08-87ca-7de94cdb7afa}|ConvertLib.dll;"
+		Debug.AspNetCompiler.VirtualPath = "/localhost_57492"
+		Debug.AspNetCompiler.PhysicalPath = "Website\"
+		Debug.AspNetCompiler.TargetPath = "PrecompiledWeb\localhost_57492\"
 		Debug.AspNetCompiler.Updateable = "true"
 		Debug.AspNetCompiler.ForceOverwrite = "true"
 		Debug.AspNetCompiler.FixedNames = "false"
 		Debug.AspNetCompiler.Debug = "True"
-		Release.AspNetCompiler.VirtualPath = "/Website"
-		Release.AspNetCompiler.PhysicalPath = "..\Website\"
-		Release.AspNetCompiler.TargetPath = "PrecompiledWeb\Website\"
+		Release.AspNetCompiler.VirtualPath = "/localhost_57492"
+		Release.AspNetCompiler.PhysicalPath = "Website\"
+		Release.AspNetCompiler.TargetPath = "PrecompiledWeb\localhost_57492\"
 		Release.AspNetCompiler.Updateable = "true"
 		Release.AspNetCompiler.ForceOverwrite = "true"
 		Release.AspNetCompiler.FixedNames = "false"
 		Release.AspNetCompiler.Debug = "False"
-		VWDPort = "1086"
+		SlnRelativePath = "Website\"
+		DefaultWebSiteLanguage = "Visual C#"
 	EndProjectSection
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "ConvertLib", "Converter\ConvertLib.csproj", "{D3173D35-982E-4A08-87CA-7DE94CDB7AFA}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|.NET = Debug|.NET
+		Debug|Any CPU = Debug|Any CPU
+		Debug|Mixed Platforms = Debug|Mixed Platforms
+		Release|.NET = Release|.NET
+		Release|Any CPU = Release|Any CPU
+		Release|Mixed Platforms = Release|Mixed Platforms
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{838EA5E3-8A94-4D2B-AF1D-121BF7AA2B0E}.Debug|.NET.ActiveCfg = Debug|.NET
-		{838EA5E3-8A94-4D2B-AF1D-121BF7AA2B0E}.Debug|.NET.Build.0 = Debug|.NET
+		{E7ABC5AC-CB11-48B0-A07E-D36D0F5DB7AE}.Debug|.NET.ActiveCfg = Debug|Any CPU
+		{E7ABC5AC-CB11-48B0-A07E-D36D0F5DB7AE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E7ABC5AC-CB11-48B0-A07E-D36D0F5DB7AE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E7ABC5AC-CB11-48B0-A07E-D36D0F5DB7AE}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{E7ABC5AC-CB11-48B0-A07E-D36D0F5DB7AE}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
+		{E7ABC5AC-CB11-48B0-A07E-D36D0F5DB7AE}.Release|.NET.ActiveCfg = Debug|Any CPU
+		{E7ABC5AC-CB11-48B0-A07E-D36D0F5DB7AE}.Release|Any CPU.ActiveCfg = Debug|Any CPU
+		{E7ABC5AC-CB11-48B0-A07E-D36D0F5DB7AE}.Release|Any CPU.Build.0 = Debug|Any CPU
+		{E7ABC5AC-CB11-48B0-A07E-D36D0F5DB7AE}.Release|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{E7ABC5AC-CB11-48B0-A07E-D36D0F5DB7AE}.Release|Mixed Platforms.Build.0 = Debug|Any CPU
+		{D3173D35-982E-4A08-87CA-7DE94CDB7AFA}.Debug|.NET.ActiveCfg = Debug|Any CPU
+		{D3173D35-982E-4A08-87CA-7DE94CDB7AFA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D3173D35-982E-4A08-87CA-7DE94CDB7AFA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D3173D35-982E-4A08-87CA-7DE94CDB7AFA}.Debug|Mixed Platforms.ActiveCfg = Debug|Any CPU
+		{D3173D35-982E-4A08-87CA-7DE94CDB7AFA}.Debug|Mixed Platforms.Build.0 = Debug|Any CPU
+		{D3173D35-982E-4A08-87CA-7DE94CDB7AFA}.Release|.NET.ActiveCfg = Release|Any CPU
+		{D3173D35-982E-4A08-87CA-7DE94CDB7AFA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D3173D35-982E-4A08-87CA-7DE94CDB7AFA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D3173D35-982E-4A08-87CA-7DE94CDB7AFA}.Release|Mixed Platforms.ActiveCfg = Release|Any CPU
+		{D3173D35-982E-4A08-87CA-7DE94CDB7AFA}.Release|Mixed Platforms.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
