--- src/dotnet/Test.asmx
@@ -1,1 +1,0 @@
-﻿<%@ WebService Language="C#" CodeBehind="~/App_Code/Test.cs" Class="TestService" %>
