--- src/dotnet/Web.config
@@ -1,42 +1,0 @@
-﻿<?xml version="1.0"?>
-<!-- 
-    Note: As an alternative to hand editing this file you can use the 
-    web admin tool to configure settings for your application. Use
-    the Website->Asp.Net Configuration option in Visual Studio.
-    A full list of settings and comments can be found in 
-    machine.config.comments usually located in 
-    \Windows\Microsoft.Net\Framework\v2.x\Config 
--->
-<configuration>
-	<appSettings>
-		<!-- <add key="OutputMimetype" value="text/plain"/> -->
-	</appSettings>
-	<connectionStrings/>
-	<system.web>
-		<compilation debug="true"/>
-		<authentication mode="Forms">
-			<forms loginUrl="SignIn.aspx"/>
-		</authentication>
-		<pages>
-			<controls>
-				<add tagPrefix="asp" namespace="System.Web.UI" assembly="System.Web.Extensions, Version=3.5.0.0, Culture=neutral, PublicKeyToken=31BF3856AD364E35"/>
-				<add tagPrefix="asp" namespace="System.Web.UI.WebControls" assembly="System.Web.Extensions, Version=3.5.0.0, Culture=neutral, PublicKeyToken=31BF3856AD364E35"/>
-			</controls>
-		</pages>
-		<httpHandlers>
-			<remove verb="*" path="*.asmx"/>
-			<add verb="*" path="*.asmx" validate="false" type="System.Web.Script.Services.ScriptHandlerFactory, System.Web.Extensions, Version=3.5.0.0, Culture=neutral, PublicKeyToken=31BF3856AD364E35"/>
-			<add verb="*" path="*_AppService.axd" validate="false" type="System.Web.Script.Services.ScriptHandlerFactory, System.Web.Extensions, Version=3.5.0.0, Culture=neutral, PublicKeyToken=31BF3856AD364E35"/>
-		</httpHandlers>
-		<httpModules>
-		</httpModules>
-		<webServices>
-			<protocols>
-				<add name="HttpGet"/>
-				<add name="HttpPost"/>
-				<remove name="HttpSoap12"/>
-			</protocols>
-		</webServices>
-	        <customErrors mode="Off"/>
-	</system.web>
-</configuration>
