+++ src/dotnet/Website/assets/images/generate-button.png
