+++ src/dotnet/Website/assets/images/icon.png
